Create TaskList, TextUi, Storage, Parser classes and packaging them into duke/task packages
