+++ src/main/java/tasks/Deadline.java
@@ -1,43 +1,44 @@
+package tasks;
+import duke.*;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
+public class Deadline extends Task {
 
-public class Event extends Task {
-
-    protected LocalDate at;
-    protected LocalDateTime atTime;
+    protected LocalDate by;
+    protected LocalDateTime byTime;
     private static final DateTimeFormatter dateFormat = DateTimeFormatter.ofPattern("dd-MM-yyyy");
     private static final DateTimeFormatter dateTimeFormat = DateTimeFormatter.ofPattern("dd-MM-yyyy HH:mm");
 
-
-    public Event(String description, String at) throws DukeException {
+    public Deadline(String description, String by) throws DukeException {
         super(description);
-        if (isDateFormat(at)) {
-            this.at = LocalDate.parse(at, dateFormat);
-        } else if (isDateTimeFormat(at)) {
-            this.atTime = LocalDateTime.parse(at, dateTimeFormat);
+        if (isDateFormat(by)) {
+            this.by = LocalDate.parse(by, dateFormat);
+        } else if (isDateTimeFormat(by)) {
+            this.byTime = LocalDateTime.parse(by, dateTimeFormat);
         } else {
             throw new DukeException("The date format parsed is incorrect! It should be dd-MM-yyyy or dd-MM-yyyy HH:mm!");
         }
     }
 
+
     @Override
     public String toFileFormat() {
-        if (at != null) {
-            return "E" + super.toFileFormat() + "," + formatDate(at);
+        if (by != null) {
+            return "D" + super.toFileFormat() + "," + formatDate(by);
         } else {
-            return "E" + super.toFileFormat() + "," + formatDateTime(atTime);
+            return "D" + super.toFileFormat() + "," + formatDateTime(byTime);
         }
     }
 
     @Override
     public String toString() {
-        if (this.at != null) {
-            return "[E]" + super.toString() + " (at: " + formatDate(at) + ")";
+        if (this.by != null) {
+            return "[D]" + super.toString() + " (by: " + formatDate(by) + ")";
         } else {
-            return "[E]" + super.toString() + " (at: " + formatDateTime(atTime) + ")";
+            return "[D]" + super.toString() + " (by: " + formatDateTime(byTime) + ")";
         }
     }
 
@@ -45,23 +46,21 @@ public class Event extends Task {
         try {
             LocalDateTime.parse(dateTimeString, dateTimeFormat);
             return true;
-        } catch (DateTimeParseException e) {
+        } catch (Exception e) {
             System.out.println("Checking DateTime Format!");
         }
         return false;
     }
 
-
-    private boolean isDateFormat(String dateString) {
+    private boolean isDateFormat(String dateString)  {
         try {
             LocalDate.parse(dateString, dateFormat);
             return true;
-        } catch (DateTimeParseException e) {
+        } catch (Exception e) {
             System.out.println("Checking Date Format!");
         }
         return false;
     }
-
 
     public String formatDateTime(LocalDateTime dateTime) {
         String day = dateTime.getDayOfMonth() < 10 ? "0" + dateTime.getDayOfMonth() : "" + dateTime.getDayOfMonth();
@@ -80,3 +79,4 @@ public class Event extends Task {
     }
 
 }
+
